moved SplitPCAP and attendant utility file to org.apache.nifi.processors.network.util
