+++ nifi-extension-bundles/nifi-network-bundle/nifi-network-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 org.apache.nifi.processors.network.ParseNetflowv5
-org.apache.nifi.processors.network.SplitPCAP
+org.apache.nifi.processors.network.pcap.SplitPCAP
